correcao para outro editor sem ser o  gutenberg ou o classico

+++ app/Posts.php
@@ -137,7 +137,7 @@ Class Posts {
 
 		if($delete_old){
 			$delete_old = Helpers::getOption('path_rlout').'/'.$delete_old;
-			if($delete_old!=$dir_base || !strpos($dir_base, $_POST['post_name'].'/')){
+			if($delete_old!=$dir_base || (!strpos($dir_base, $_POST['post_name'].'/') && !empty($_POST['post_name']))){
 				Helpers::rrmdir($delete_old);
 				S3::remove_file($delete_old . 'index.html');
 			}
